Figured out the problem
When the numbers got to three digits the indexes were not working
properly and would repeat in a unknown pattern.

I set the Champernowne.current to 100 and looked at a small set of data
and started to rewrite the iterator loop. Since the iterator always
requires a result, and is considered terminated when a `None` result is
returned, I realized that I needed to add a function to make this
easier.

This solution does not feel very safe to me. It uses unbounded
recursion. Though I don't think it's possible to get stuck in a loop of
ever returning None values, if it were it would cause a deadlock. I
don't think that's something that I need to worry about for this case,
but it is something I thought about.