Remove the option value
I don't remember why I have this here. It's possibly a remnant to how I
was creating the struct beforehand, or maybe it made sense to
differentiate between having a index of zero and needing to start on the
next numbers character set.

In any case, with the recursive setup it's not required anymore and it
looks a lot cleaner without it.